perf(ui, client): remove testId from FormSection component
