+++ packages/client/src/utility/catalog.ts
@@ -1,4 +1,32 @@
+import { runPixel } from "@semoss/sdk/react";
 import type { Engine, Project } from "@semoss/shared";
+import { z } from "@semoss/ui/next";
+
+/** Catalog names (projects and engines alike) may only contain alphanumerics, dashes, and spaces. */
+export const CATALOG_NAME_PATTERN = /^[\w\-\s]+$/;
+
+/**
+ * Check whether a catalog name is already taken (CheckEngineName pixel).
+ */
+export const isEngineNameTaken = async (name: string): Promise<boolean> => {
+	const response = await runPixel<[{ exists: boolean }]>(
+		`CheckEngineName ( "${name}") ;`,
+	);
+	return Boolean(response.pixelReturn[0]?.output?.exists);
+};
+
+/** Shared zod schema for a catalog entry's (project or engine) name field. */
+export const catalogNameSchema = z
+	.string()
+	.min(1, "Catalog name is required")
+	.regex(
+		CATALOG_NAME_PATTERN,
+		"Catalog names can only contain alphanumeric characters and dashes.",
+	)
+	.refine(async (name) => !(await isEngineNameTaken(name)), {
+		message: "This Catalog name has already been used, please try another.",
+	});
+
 /**
  * Utility to check if it is a project type
  * @param type
